Added example of how to add a custom link to admin course listing
When viewing the list of courses in the Canvas Admin, quicklinks appear
as you mouseover
each course.  These quicklinks include 'Settings', 'Statistics', and
'Homepage'.  This
javascript will add 'delete' to this list.  When learning canvas, and
specifically when
working in a sandbox environment, it can be useful to be able to
quickly delete a course
without having to visit the course homepage or the settings page.